refactor(slack): extract channel-management actions to clear lint ratchets
actions.ts exceeded the 700-line max-lines ratchet after the five
management functions landed. Move them into channel-management-actions.ts
and re-export from actions.ts so the lazy runtime module and its typed
key surface stay unchanged. Rewrite the message-tool-api.ts management
filter as a type-safe loop (matching the sibling id-actions pattern) to
drop the two uncommented type assertions that broke the
assertion-safety ratchet baseline.
