+++ extensions/slack/src/actions.ts
@@ -214,7 +214,7 @@ function hasSlackPlatformError(err: unknown, code: string): boolean {
   return (data as { error?: unknown }).error === code;
 }
 
-async function getClient(opts: SlackActionClientOpts = {}, mode: "read" | "write" = "read") {
+export async function getClient(opts: SlackActionClientOpts = {}, mode: "read" | "write" = "read") {
   if (opts.client) {
     return opts.client;
   }
@@ -598,62 +598,6 @@ export async function listSlackPins(
   return (result.items ?? []) as SlackPin[];
 }
 
-export type SlackChannelManagementResult = {
-  channelId?: string;
-  name?: string;
-};
-
-export async function createSlackChannel(
-  name: string,
-  opts: SlackActionClientOpts & { isPrivate?: boolean } = {},
-): Promise<SlackChannelManagementResult> {
-  const client = await getClient(opts, "write");
-  const result = await client.conversations.create({
-    name,
-    ...(opts.isPrivate ? { is_private: true } : {}),
-  });
-  return { channelId: result.channel?.id, name: result.channel?.name };
-}
-
-export async function renameSlackChannel(
-  channelId: string,
-  name: string,
-  opts: SlackActionClientOpts = {},
-): Promise<SlackChannelManagementResult> {
-  const client = await getClient(opts, "write");
-  const result = await client.conversations.rename({ channel: channelId, name });
-  return { channelId: result.channel?.id, name: result.channel?.name };
-}
-
-export async function addSlackChannelMember(
-  channelId: string,
-  userId: string,
-  opts: SlackActionClientOpts = {},
-): Promise<SlackChannelManagementResult> {
-  const client = await getClient(opts, "write");
-  await client.conversations.invite({ channel: channelId, users: userId });
-  return { channelId };
-}
-
-export async function removeSlackChannelMember(
-  channelId: string,
-  userId: string,
-  opts: SlackActionClientOpts = {},
-): Promise<SlackChannelManagementResult> {
-  const client = await getClient(opts, "write");
-  await client.conversations.kick({ channel: channelId, user: userId });
-  return { channelId };
-}
-
-export async function archiveSlackChannel(
-  channelId: string,
-  opts: SlackActionClientOpts = {},
-): Promise<SlackChannelManagementResult> {
-  const client = await getClient(opts, "write");
-  await client.conversations.archive({ channel: channelId });
-  return { channelId };
-}
-
 type SlackFileInfoSummary = {
   id?: string;
   name?: string;
@@ -800,3 +744,12 @@ export async function downloadSlackFile(
 
   return results?.[0] ?? null;
 }
+
+export {
+  addSlackChannelMember,
+  archiveSlackChannel,
+  createSlackChannel,
+  removeSlackChannelMember,
+  renameSlackChannel,
+  type SlackChannelManagementResult,
+} from "./channel-management-actions.js";
